# contracts/quality-api/openapi.yaml
openapi: 3.1.0
...
info:
  title: CaseLoop Quality API
  version: "2.0.0"
  summary: AI 应用质量自治底座——被治理应用的质量契约（v2 强化版）
  description: |
    Quality API v2（plan-v3 §2.3.1 / T0）。任何 LLM 应用实现本契约即可被 CaseLoop 纳管。

    核心不变量：
    - **VersionSet 不可变**：{prompt, KB manifest, model+params} 创建后内容不可改，
      变更 = 新 VersionSet（新 id + 新 digest）；只有生命周期状态可迁移。
    - **生命周期**：draft → stage → canary → promote | rollback → status。
    - **写面纪律**：全部写操作要求 CAS（If-Match 头 或 body expected_revision）、
      Idempotency-Key 头，且为异步——返回 operation_id，经 GET /v2/operations/{id} 查结果。
    - **写面仅 Release Controller 可调**：quality:write scope 令牌只签发给
      Release Controller（控制面组件，非 LLM）。应用自身、Agent 均不得持有。
    - **读面**：GET /v2/logs、GET /v2/feedback（quality:read 或 quality:write）。

    digest 规则：所有 digest = sha256(JCS(RFC 8785) canonical JSON)，
    表示为 `sha256:<64 hex>`；VersionSet 的完整版本 digest 覆盖其
    {prompt, kb_manifest, model} 三元组内容。

  contact:
    name: CaseLoop contracts
    url: https://caseloop.dev/contracts
servers:
  - url: http://127.0.0.1:8080
    description: demo-app 本地默认（conformance suite 的 BASE_URL 缺省值）
...
paths:
  /v2/versionsets:
    post:
      tags: [versionsets]
      summary: 创建新 VersionSet（初始状态 draft）
      description: |
        内容创建即不可变。digest 由服务端按 JCS+SHA-256 计算并返回；
        幂等：相同 Idempotency-Key + 相同内容 → 返回同一 VersionSet（不新建）。
      operationId: createVersionSet
      security:
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VersionSetContent"
      responses:
        "201":
          description: 创建成功（同步创建资源本身；生命周期迁移全部走异步）
          headers:
            ETag:
              description: 当前 revision（CAS 用）
              schema: { type: string, example: "\"1\"" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
    get:
      tags: [versionsets]
      summary: 列出 VersionSet
      operationId: listVersionSets
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LifecycleStatus"
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 分页列表
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionSetPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v2/versionsets/{versionsetId}:
    get:
      tags: [versionsets]
      summary: 读取单个 VersionSet（含完整 digest 与生命周期状态）
      operationId: getVersionSet
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
      responses:
        "200":
          description: 成功
          headers:
            ETag:
              description: 当前 revision，写操作 If-Match 用它
              schema: { type: string, example: "\"3\"" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionSet"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v2/versionsets/{versionsetId}/status:
    get:
      tags: [lifecycle]
      summary: 生命周期状态查询（含灰度信息与迁移历史）
      operationId: getVersionSetStatus
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
      responses:
        "200":
          description: 成功
          headers:
            ETag:
              schema: { type: string, example: "\"3\"" }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VersionSetStatus"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /v2/versionsets/{versionsetId}/evaluate:
    post:
      tags: [versionsets]
      summary: 对精确不可变 VersionSet 执行一条候选评测探针
      description: |
        Gate Runner 的内部只读执行面。该调用不改变生命周期状态，
        也不使用当前 active 版本或故障覆盖；它必须从指定 VersionSet
        的 prompt、KB manifest 及 model digest 精确重建运行配置。
        任何注册资产缺失或 digest 漂移都返回 409，不得回退到 active。
      operationId: evaluateVersionSetCandidate
      x-internal: true
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [message]
              properties:
                message: { type: string, minLength: 1, maxLength: 2000 }
                session_id: { type: [string, "null"] }
                user_ref: { type: [string, "null"] }
      responses:
        "200":
          description: 真实 provider 候选回复，并回显实际执行的组件 digest
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                required:
                  - request_id
                  - answer
                  - status
                  - versionset_id
                  - prompt_digest
                  - kb_manifest_digest
                  - model_digest
                  - retrieval
                properties:
                  request_id: { type: string }
                  answer: { type: string }
                  status: { type: string, enum: [ok, provider_error] }
                  versionset_id: { type: string }
                  prompt_digest: { $ref: "#/components/schemas/Digest" }
                  kb_manifest_digest: { $ref: "#/components/schemas/Digest" }
                  model_digest: { $ref: "#/components/schemas/Digest" }
                  trace_id: { type: [string, "null"] }
                  retrieval: { type: array, items: { type: object } }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /v2/versionsets/{versionsetId}/stage:
    post:
      tags: [lifecycle]
      summary: draft → staged（进入待灰度区）
      operationId: stageVersionSet
      security:
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LifecycleRequest"
      responses:
        "202":
          $ref: "#/components/responses/AcceptedOperation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RevisionConflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v2/versionsets/{versionsetId}/canary:
    post:
      tags: [lifecycle]
      summary: staged → canary（按百分比灰度）
      operationId: canaryVersionSet
      security:
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CanaryRequest"
      responses:
        "202":
          $ref: "#/components/responses/AcceptedOperation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RevisionConflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v2/versionsets/{versionsetId}/promote:
    post:
      tags: [lifecycle]
      summary: canary → active（全量；原 active 自动转 superseded）
      operationId: promoteVersionSet
      security:
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromoteRequest"
      responses:
        "202":
          $ref: "#/components/responses/AcceptedOperation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RevisionConflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v2/versionsets/{versionsetId}/rollback:
    post:
      tags: [lifecycle]
      summary: 回滚：canary/active → rolled_back，目标版本保持或恢复为 active
      description: |
        回滚是语义级操作而非内容修改：本 VersionSet 转 rolled_back。
        canary 回滚保留当前 active 基线；active 回滚则将 rollback_to 指定的
        （或 previous = 上一个 active）VersionSet 恢复为 active。
        回滚失败（下游不一致）时 operation 转 failed 且必须触发人工接管事件
        （见 contracts/events/state-machines.yaml Release 状态机）。
      operationId: rollbackVersionSet
      security:
        - quality_oauth: [quality:write]
      parameters:
        - $ref: "#/components/parameters/VersionSetId"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RollbackRequest"
      responses:
        "202":
          $ref: "#/components/responses/AcceptedOperation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/RevisionConflict"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/ValidationFailed"
  /v2/operations/{operationId}:
    get:
      tags: [operations]
      summary: 查询异步写操作结果
      description: |
        写操作（stage/canary/promote/rollback）均为异步：先返 202 + operation_id，
        本端点轮询至 succeeded/failed。operation 记录有 TTL（【待定】默认 24h），
        过期清理后查询返回 410 operation_expired——发布侧 UNKNOWN 语义见此：
        调用方查不到终态时必须按 Release 状态机走 reconcile，不得盲目重试。
      operationId: getOperation
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - name: operationId
          in: path
          required: true
          schema:
            type: string
            pattern: "^op_[0-9A-Za-z]{8,64}$"
      responses:
        "200":
          description: 操作当前状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "410":
          $ref: "#/components/responses/OperationExpired"
  /v2/logs:
    get:
      tags: [read]
      summary: 运行日志读面（plan-v3 保留接口）
      description: |
        每次请求一条：携带实际生效的 prompt/kb/model digest，
        使控制面可以把 badcase 精确绑定到版本。【待定】分页/过滤字段
        plan-v3 未细化，当前定为 cursor 分页 + 时间窗 + versionset_id 过滤。
      operationId: getLogs
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: versionset_id
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 日志页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v2/feedback:
    get:
      tags: [read]
      summary: 用户反馈读面（plan-v3 保留接口）
      operationId: getFeedback
      security:
        - quality_oauth: [quality:read]
        - quality_oauth: [quality:write]
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date-time }
        - name: to
          in: query
          schema: { type: string, format: date-time }
        - name: versionset_id
          in: query
          schema: { type: string }
        - name: rating
          in: query
          schema: { type: string, enum: [positive, negative, neutral] }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        "200":
          description: 反馈页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    quality_oauth:
      type: oauth2
      description: |
        OAuth2 client credentials。scope 语义：
        - **quality:read**：读面（GET logs/feedback/versionsets/operations）。
        - **quality:write**：写面（创建 VersionSet、全部生命周期迁移、admin 注入）。
          **写 scope 令牌仅签发给 Release Controller**（plan-v3 §2.3.1 硬约束）；
          持有 quality:write 的调用方同时视为有读权限。
      flows:
        clientCredentials:
          tokenUrl: /oauth/token
          scopes:
            quality:read: "读面：logs/feedback/versionsets/operations 查询"
            quality:write: "写面：仅 Release Controller 可持有"
  parameters:
    VersionSetId:
      name: versionsetId
      in: path
      required: true
      schema:
        type: string
        pattern: "^vs_[0-9A-Za-z]{8,64}$"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        幂等键（调用方生成，建议 UUID）。相同 key + 相同请求指纹的重放
        必须返回同一个 operation（不重复执行）；相同 key 但请求体不同 →
        422 validation_failed（details.subcode=idempotency_key_conflict）。
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
...
    Unauthorized:
      description: 未认证（令牌缺失/无效）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: unauthorized, message: "missing or invalid token" }
    Forbidden:
      description: 已认证但 scope 不足（如持 quality:read 调写面）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: forbidden, message: "scope quality:write required; write surface is restricted to Release Controller" }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: not_found, message: "versionset not found" }
...
    ValidationFailed:
      description: 422 validation_failed——请求体验证失败或非法生命周期迁移（details.subcode 区分：illegal_transition / schema_invalid / idempotency_key_conflict）
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: validation_failed
              message: "illegal transition: cannot canary from status draft"
              details: { subcode: illegal_transition, current_status: draft, attempted: canary }
...
  schemas:
...
    VersionSetContent:
      type: object
      description: VersionSet 不可变内容三元组；完整版本 digest 覆盖本对象整体
      additionalProperties: false
      required: [prompt, kb_manifest, model]
      properties:
        prompt:
          $ref: "#/components/schemas/Prompt"
        kb_manifest:
          $ref: "#/components/schemas/KBManifest"
        model:
          $ref: "#/components/schemas/ModelSpec"
    LifecycleStatus:
      type: string
      description: |
        生命周期状态机（非法迁移一律 422 validation_failed）：
        draft → staged → canary → active（promote）；canary/active → rolled_back（rollback）；
        promote 时原 active 自动 → superseded。
      enum: [draft, staged, canary, active, superseded, rolled_back]
    VersionSet:
      type: object
      additionalProperties: false
      required: [versionset_id, revision, status, content, digest, created_at]
      properties:
        versionset_id:
          type: string
          pattern: "^vs_[0-9A-Za-z]{8,64}$"
        revision:
          type: integer
          minimum: 1
          description: 单调递增的乐观并发 revision（ETag）；每次生命周期迁移 +1
        status:
          $ref: "#/components/schemas/LifecycleStatus"
        content:
          $ref: "#/components/schemas/VersionSetContent"
        digest:
          type: string
          description: 完整版本 digest = sha256(JCS(content))
          $ref: "#/components/schemas/Digest"
        created_at:
          type: string
          format: date-time
        labels:
          type: object
          additionalProperties: { type: string }
    VersionSetPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VersionSet"
        next_cursor:
          type: string
          description: 无更多页时缺省
...
security: []